Improve flexibility of leapfrog
This starts using PRE_ and POST_ scripts, to allow to tap into
the process to leapfrog RPC.

This should still do an OSA leapfrog, but shows the ability to
load the PRE_ and POST_ scripts.

It also gives an example of what these could be (in comments),
for later iterations.