LA-321 Add support for generating test resources
The script `generate_resources.py` is added to provide basic support for
adding resources, e.g. compute instances or block storage volumes, to
allow the impact of an upgrade, from an end-users perspective, to be
measured.

The script assumes the openstack-ansible playbook
`os-tempest-install.yml` has already been run on the deployment.

The playbook `generate-resources.yml` ensures the necessary Pip
dependencies are installed prior to running `generate_resources.py`
using a virtualenv.

An example of the playbook's use is:

openstack-ansible \
    -e server_count=10 \
    -e volume_count=25 \
    -e network_count=12 \
    generate-resources.yml