RLM-554 Various MaaS fixes
Restores playbook to remove old neutron container maas entries
which was removed in RLM-416.

Runs maas-install before and after leapfrog which gets rpc-maas
if it's not present, installs, and verifies.  It also handles
the jumping around of the maas_release variable and sets it to
master by default for now.