Added QC plays/roles/tasks for pre/post upgrade QC
In order to automate the pre/post upgrade QC process we needed
to identify the different QC mechanisms across the different openstack
service levels. This commit focuses on 4; environment, infrastructure,
openstack, and swift. The individual tasks check the status of the services
and output them to a file for each for a review. Each level can be called
individually as needed for whatever step (pre/post) that the user desires