RE-1162 Remove legacy .travis.yml file
This commit removes the .travis.yml, and also passes a custom workdir
to tox to avoid failures where files in tox's venvs have excessively
long shebang lines.  At present, we're hitting such an issue with [1]
which is causing the job to fail in the most cryptic way possible.

Thanks to @git-harry for the fix.

[1] PR_rpc-upgrades-master-container-ansible-syntax-tox-test