Add var to override running of upgrade in tests
Currently we upgrade no matter what the user is trying to in the
tests directory of scripts. This commit adds an override variable
to allow the tester to choose whether to upgrade or not at run time
of the build.sh script. This will allow the user to have more granular
control when testing