RLM-1291 Adds hook in for major upgrades
Currently fires off the test-leapfrog.sh script for the major upgrade
as the leapfrog jump from Mitaka to Newton appears to work.

There are a few tweaks that will need to be done around the config
file migrations migrations that will need to be done to get it
fully working.