ASC-564 Persist release envvars for test results
In order to capture accurate metadata about Molecule system tests, the
"RPC_RELEASE" and "RPC_PRODUCT_RELEASE" environment variables need to be
stored in the "MNAIO_VAR_FILE". Values in this file are re-sourced by the
`run` script to ensure they are used by the system tests in reporting
their results.