RE-1590 Install required packages before test
Prior to executing tests there are some pre-requisite packages
which need to be installed for various reasons.

This patch adds some new packages required to install for tests
when using nodepool nodes, and consolidates some separated
package installs into a single action to ensure that the package
manager is able to resolve everything at once. Also, the apt
update is split from the apt install as using the && to combine
them has been shown upstream in OSA to cause unexpected
side-effects which we should rather avoid.