RE-1590 Exclude tmpfs/iso9660 from free space check
The amount of free space available for (dev)tmpfs file systems
such as /run, /sys, etc are not important, nor are read-only
file systems like iso9660 which get used for config drive in
cloud systems (used for testing), and for DVD/CD removable
media drives. As such, we skip those from the check for free
space.

We also ensure that the type is output when giving an error
so that it provides a more informed output.