FLEEK-151 Fix enable_lxc_fs_check flag
Previously if the flag was set to disable it would skip the
task to gather the data, but the following task would fail because
that data was missing.  This sets it to always gather that
information, tosses aninformational alert and skips the hard failure.

Also touches preflight completed file at end of preflight script
to allow the script to be ran independently ahead of the upgrade
so that options can be passed to skip items if needed.