Adds additional reports for checking status
Checks list of endpoints to ensure lb addreses match
appropriate endpoints and lists in file.

Adds osa_health role for gathering data from all hosts
and generates a TLDR report that can be easily parsed.
The role currently generates a list of updates that need
to be done on all containers/hosts and does a check for
zombie processes.

Renames status file to openstack-status.